requirements.txt
Django>=4.2,<5.0mysqlclient>=2.2.0python-decouple>=3.8django-environ>=0.11.2django-admin-interface>=0.25.0djangorestframework